Python/requirements.txt
requests
ffmpeg-python